requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
appscript==1.3.0
asttokens==3.0.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
docx2pdf==0.1.8
executing==2.2.0
Faker==35.2.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain-core==0.3.33
langchain-ollama==0.2.3
langchain-openai==0.3.3
langsmith==0.3.3
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
ollama==0.4.7
openai==1.60.2
orjson==3.10.15
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11
PyMuPDF==1.25.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0
wcwidth==0.2.13
zstandard==0.23.0
